Heuristics + new getter on Config.h
Heuristics 100% done, next update i'll try to implement A* and test it.
Added a "getter" on Config.h to get access to the matrix board in order to implement the Heuristics.
Included two heuristics: Number of misplaced tiles & Manhattan Distance (Teacher asked only one.)
Feel free to optimize the code. Couldn't find the bug on the print of the Path. The code is in an infinite loop, because the nodes always have the same parent instead of the Node in the last depth.